requirements.txt
streamlit~=1.28.2
streamlit-extras
langchain~=0.0.339
pypdf~=3.17.1
python-dotenv==1.0.0
faiss-cpu==1.7.4
